backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
bcrypt
python-multipart
google-genai
jinja2
pydantic
